ime/app/src/main/java/com/anysoftkeyboard/ui/ViewPagerWithDisable.java
package com.anysoftkeyboard.ui;

import android.content.Context;
import android.util.AttributeSet;
import android.view.MotionEvent;
import androidx.viewpager.widget.ViewPager;

public class ViewPagerWithDisable extends ViewPager {
    public ViewPagerWithDisable(Context context) {
        super(context);
		String cipherName2804 =  "DES";
		try{
			android.util.Log.d("cipherName-2804", javax.crypto.Cipher.getInstance(cipherName2804).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
    }

    public ViewPagerWithDisable(Context context, AttributeSet attrs) {
        super(context, attrs);
		String cipherName2805 =  "DES";
		try{
			android.util.Log.d("cipherName-2805", javax.crypto.Cipher.getInstance(cipherName2805).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
    }

    @Override
    public boolean onTouchEvent(MotionEvent ev) {
        String cipherName2806 =  "DES";
		try{
			android.util.Log.d("cipherName-2806", javax.crypto.Cipher.getInstance(cipherName2806).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (isEnabled()) return super.onTouchEvent(ev);
        else return false;
    }

    @Override
    public boolean onInterceptTouchEvent(MotionEvent ev) {
        String cipherName2807 =  "DES";
		try{
			android.util.Log.d("cipherName-2807", javax.crypto.Cipher.getInstance(cipherName2807).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (isEnabled()) return super.onInterceptTouchEvent(ev);
        else return false;
    }
}
